Add datetime to Redux
