+++ v2/frontend/app/clientLayout.tsx
@@ -1,8 +1,6 @@
 "use client"
 
 import { orange } from "@mui/material/colors";
-import Divider from "@mui/material/Divider";
-import Drawer from "@mui/material/Drawer";
 import { createTheme } from "@mui/material/styles";
 import ThemeProvider from "@mui/system/ThemeProvider";
 import React from "react";
@@ -10,10 +8,8 @@ import { Provider as ReduxProvider } from 'react-redux'
 
 import NavBar from "../components/NavBar";
 import SearchModal from "../components/SearchModal";
-import { selectCurrentBuilding, setCurrentBuilding } from "../redux/currentBuildingSlice";
-import { useDispatch, useSelector } from "../redux/hooks";
 import store from '../redux/store'
-import BuildingInfo from "../views/BuildingInfo";
+import BuildingDrawer from "../views/BuildingDrawer";
 
 const theme = createTheme({
   palette: {
@@ -22,8 +18,6 @@ const theme = createTheme({
     },
   },
 });
-
-const drawerWidth = 400;
 
 const ClientLayout: React.FC<{
   children: React.ReactNode;
@@ -46,37 +40,13 @@ const App: React.FC<{
   children: React.ReactNode;
 }> = ({ children }) => {
   const [searchOpen, setSearchOpen] = React.useState(false);
-  const dispatch = useDispatch();
-  const currentBuilding = useSelector(selectCurrentBuilding);
-  const drawerOpen = !!currentBuilding;
 
   return (
     <>
       <NavBar setSearchOpen={setSearchOpen}/>
       <SearchModal open={searchOpen} setOpen={setSearchOpen}/>
       {children}
-      <Drawer
-        sx={{
-          width: drawerWidth,
-          flexShrink: 0,
-          "& .MuiDrawer-paper": {
-            width: drawerWidth,
-            boxSizing: "border-box",
-          },
-        }}
-        variant="persistent"
-        anchor="right"
-        open={drawerOpen}
-      >
-        <Divider />
-        <BuildingInfo
-          building={currentBuilding}
-          onClose={() => dispatch(setCurrentBuilding(null))}
-          datetime={new Date()}
-          setDatetime={() => {}}
-          roomStatusData={undefined}
-        />
-      </Drawer>
+      <BuildingDrawer/>
     </>
   )
 }
